aws_lambda_functions/search_papers/requirements.txt
farm-haystack[preprocessing,faiss,inference,file-conversion,pdf]
awslambdaric
boto3